수정: 전시, 댓글 기능 관련 view, controller, dao 등 수정

+++ src/controller/WebVR/DeleteCommentController.java
@@ -4,6 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import controller.Controller;
+import controller.WebVR.user.UserSessionUtils;
 import model.dao.CommentDAO;
 
 public class DeleteCommentController implements Controller {
@@ -12,14 +13,19 @@ public class DeleteCommentController implements Controller {
 
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
-
+		// 로그인 여부
+		if (!UserSessionUtils.hasLogined(request.getSession())) { // 로그인 안되어있는 있는 경우
+			return "redirect:/WebVR/login/form";	
+		}
+		
 		String artworkId = request.getParameter("artworkId");
+		System.out.println("artworkId: " + artworkId);
 		
 		int commentId = Integer.parseInt(request.getParameter("commentId"));
 //		int result = commentDAO.deleteComment(commentId);
 		commentDAO.delete(commentId);
 		
-		return "redirect:/WebVR/exhb/artwork?artworkId=" + artworkId;
+		return "redirect:/WebVR/artwork/comment?artworkId=" + artworkId;
 	}
 
 }
